git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@20346 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/main/java/org/gusdb/wdk/model/jspwrap/DatasetBean.java
@@ -33,7 +33,7 @@ public class DatasetBean {
      * @return
      * @see org.gusdb.wdk.model.dataset.Dataset#getDatasetId()
      */
-    public int getDatasetId() {
+    public long getDatasetId() {
         return dataset.getDatasetId();
     }
