+++ Model/src/main/java/org/gusdb/wdk/model/user/dataset/irods/IrodsUserDatasetSession.java
@@ -6,16 +6,16 @@ import org.gusdb.wdk.model.user.dataset.UserDatasetFile;
 import org.gusdb.wdk.model.user.dataset.json.JsonUserDatasetSession;
 
 public class IrodsUserDatasetSession extends JsonUserDatasetSession {
-	
+
   public IrodsUserDatasetSession(Path usersRootDir) {
     super(new IrodsUserDatasetStoreAdaptor(), usersRootDir);
   }
-		  
+
   @Override
-  public UserDatasetFile getUserDatasetFile(Path path, Integer userDatasetId) {
+  public UserDatasetFile getUserDatasetFile(Path path, Long userDatasetId) {
     return new IrodsUserDatasetFile(path, userDatasetId);
   }
-		  
+
   @Override
   public void close() {
     ((IrodsUserDatasetStoreAdaptor) adaptor).closeSession();
